+++ macos/Flutter/GeneratedPluginRegistrant.swift
@@ -23,7 +23,7 @@ import path_provider_macos
 import photo_manager
 import screen_retriever
 import share_plus_macos
-import shared_preferences_macos
+import shared_preferences_foundation
 import sqflite
 import sqlite3_flutter_libs
 import url_launcher_macos