+++ lib/widgets/habits/habit_streaks.dart
@@ -3,59 +3,6 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:lotti/blocs/habits/habits_cubit.dart';
 import 'package:lotti/blocs/habits/habits_state.dart';
 import 'package:lotti/themes/theme.dart';
-import 'package:lotti/widgets/charts/habits/dashboard_habits_chart.dart';
-import 'package:lotti/widgets/charts/utils.dart';
-
-class HabitsStreaks extends StatelessWidget {
-  const HabitsStreaks({
-    required this.header,
-    required this.days,
-    super.key,
-  });
-
-  final String header;
-  final int days;
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<HabitsCubit, HabitsState>(
-      builder: (context, HabitsState state) {
-        final timeSpanDays = state.timeSpanDays;
-
-        final rangeStart = getStartOfDay(
-          DateTime.now().subtract(Duration(days: timeSpanDays - 1)),
-        );
-
-        final rangeEnd = getEndOfToday();
-        final showGaps = timeSpanDays < 180;
-
-        final streakItems = state.habitDefinitions.map((habitDefinition) {
-          return HabitChartLine(
-            habitDefinition: habitDefinition,
-            rangeStart: rangeStart,
-            rangeEnd: rangeEnd,
-            streakDuration: days,
-            showGaps: showGaps,
-          );
-        });
-
-        return Column(
-          children: [
-            Padding(
-              padding: const EdgeInsets.only(top: 20),
-              child: Text(
-                header,
-                style: chartTitleStyle(),
-              ),
-            ),
-            const SizedBox(height: 15),
-            ...streakItems,
-          ],
-        );
-      },
-    );
-  }
-}
 
 class HabitStreaksCounter extends StatelessWidget {
   const HabitStreaksCounter({super.key});