+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added:
+- Remove habit streaks section - not useful, streaks are apparent without
+- Chore: upgraded dependencies
+
+## [0.8.234] - 2023-01-13
 ### Changed:
 - Journal card styling
 - Entry type filter styling